requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.51.2
httpx==0.27.2
pydantic==2.9.2
PyMuPDF==1.24.10
pypdf==5.0.1
python-multipart==0.0.9
python-dotenv==1.0.1
reportlab==4.4.10
